requirements.txt
# Django itself plus addon packages. 
Django==3.1.*
channels==3.0.*
# service_identity  # is required by channels.
django-guardian==2.3.*
djangorestframework==3.12.*
drf-spectacular==0.11.*  # Online Swagger-UI for REST schema 

# Other packages required for developing.
pytest-django

